@@ -1,4 +1,4 @@
-import { Component, ChangeDetectionStrategy, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit, Input, Output, EventEmitter, ViewChild, ElementRef } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from "@angular/forms";
 import { CreateTicket } from '../../entities/write/createTicket';
 import { LOCALE } from '../../configuration/config'
@@ -12,6 +12,7 @@ import * as moment from 'moment';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class TicketCreateFormComponent implements OnInit {
+  @ViewChild("file") fileInput: ElementRef;
   @Output() onSave = new EventEmitter<CreateTicket>();
   typeOptions: Array<IOption>;
   ticketCreateForm: FormGroup;
@@ -22,7 +23,7 @@ export class TicketCreateFormComponent implements OnInit {
     this.typeOptions = [{ id: 'AAAA', title: 'TODO1' }, { id: 'BBBB', title: 'TODO2' }]       // TODO TypeOptions    
     this.ticketCreateForm = this.formBuilder.group({
       creator: [{ value: 'Brecht Carlier', disabled: true }],
-      date: [{value: moment().locale(LOCALE).format('L'), disabled: true}],
+      date: [{ value: moment().locale(LOCALE).format('L'), disabled: true }],
       title: ['', Validators.required],
       description: ['', Validators.required],
       priority: ['', Validators.required],
@@ -30,10 +31,30 @@ export class TicketCreateFormComponent implements OnInit {
     });
   }
 
-  onSubmit()
-  {
-    const formValues = this.ticketCreateForm.getRawValue();
-    this.onSave.emit(new CreateTicket(formValues.title, formValues.description, formValues.priority));
+  onfileChange(files: FileList) {
+    let validated = false;
+
+    const validedFiles = Array.from(files);
+
+    if (validedFiles.some(f => f.size > 15728640)) {
+      const file = validedFiles.find(f => f.size > 15728640);
+      alert("FILE TO BIG " + file.name);
+    }
+    else if (validedFiles.some(f => f.type !== "image/png" && f.type !== "image/jpg")) {
+      const file = validedFiles.find(f => f.type !== "image/png" && f.type !== "image/jpg");
+      alert("INCORRECT FILE " + file.name);
+    }
+    else {
+      validated = true;
+    }
+
+    if (!validated)
+      this.fileInput.nativeElement.value = "";
   }
 
+  onSubmit() {
+    const formValues = this.ticketCreateForm.getRawValue();
+    const files = this.fileInput.nativeElement.files as FileList;
+    this.onSave.emit(new CreateTicket(formValues.title, formValues.description, formValues.priority, files));
+  }
 }
