@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Response, Http } from "@angular/http";
+import { HttpClient } from "@angular/common/http";
 import { Observable } from "rxjs/Observable";
 import { READ_API } from "../../configuration/config";
 import { Ticket } from "../../entities/read/ticket";
@@ -7,10 +7,9 @@ import 'rxjs/add/operator/map';
 
 @Injectable()
 export class AllTicketsService {
-    constructor(private http: Http) { }
+    constructor(private http: HttpClient) { }
 
     all(): Observable<Array<Ticket>> {
-        return this.http.get(`${READ_API}/ticket`)
-            .map((res: Response) => res.json());
+        return this.http.get<Ticket[]>(`${READ_API}/ticket`);
     }
 }
