@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Response, Http } from "@angular/http";
+import { HttpClient } from "@angular/common/http";
 import { Observable } from "rxjs/Observable";
 import { READ_API, WRITE_API } from "../../configuration/config";
 import { TicketDetail } from "../../entities/read/ticketDetail";
@@ -7,11 +7,10 @@ import 'rxjs/add/operator/map';
 
 @Injectable()
 export class DetailTicketService {
-    constructor(private http: Http) { }
+    constructor(private http: HttpClient) { }
 
     get (id: string): Observable<TicketDetail> {
-        return this.http.get(`${READ_API}/ticket/${id}`)
-            .map((res: Response) => res.json());
+        return this.http.get<TicketDetail>(`${READ_API}/ticket/${id}`);
     }
 
     updatePriority (ticketId: string, priorityId: string){
